% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvr_annotation.R
\name{call_gene}
\alias{call_gene}
\title{Title}
\usage{
call_gene(refgene, interval = NULL, clean_cnv = NULL)
}
\arguments{
\item{refgene}{The reference genes list corresponding to your data}

\item{interval}{Could be CNV, ROH, QTL or any interval list. At least comprised by four columns Interval_ID, Chr, Start and End}

\item{clean_cnv}{The output data from cnv_clean function}
}
\value{

}
\description{
Title
}
