% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refgene.R
\name{clean_ucsc}
\alias{clean_ucsc}
\title{Clean USCS reference gene list}
\usage{
clean_ucsc(refgene)
}
\arguments{
\item{refgene}{reference gene list with UCSC version which download from UCSC website}
}
\value{
formatted gene list.
}
\description{
Clean USCS reference gene list
}
