% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_haplotype.R
\name{closer_snp}
\alias{closer_snp}
\title{Title closer_snp
Get colser snp within given interval, working for 'map' file generated from 'pre_phased function'}
\usage{
closer_snp(phased_input, chr = 14, start, end)
}
\arguments{
\item{phased_input}{map file}

\item{chr}{chromosome ID}

\item{start}{start position, unite 'bp'}

\item{end}{end position, unite 'bp'}
}
\value{
SNP position
}
\description{
Title closer_snp
Get colser snp within given interval, working for 'map' file generated from 'pre_phased function'
}
