% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_summary.R
\name{cnv_summary_plot}
\alias{cnv_summary_plot}
\title{Title
Make graphes for CNVs}
\usage{
cnv_summary_plot(clean_cnv, plot_sum_1 = NULL, plot_sum_2 = NULL)
}
\arguments{
\item{plot_sum_2}{}
}
\value{

}
\description{
Title
Make graphes for CNVs
}
