% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_visualising.R
\name{cnv_visual}
\alias{cnv_visual}
\title{Title visual CNV
Visualizing CNV on both population level and individual level}
\usage{
cnv_visual(
  clean_cnv,
  max_chr = NULL,
  chr_id = NULL,
  species = NULL,
  start_position = NULL,
  end_position = NULL,
  individual_id = NULL,
  max_chr_length = 160,
  plot_gene = NULL,
  refgene = "ARS_ens",
  plot_title = NULL,
  report_id = NULL,
  pedigree = NULL,
  show_name = c(0, 160),
  width_1 = 13,
  height_1 = 10,
  folder = "cnv_visual"
)
}
\arguments{
\item{clean_cnv}{cnv list file, standard file was generated from cnv_clean function}

\item{max_chr}{the maximum number of chromosomes to plot, it used for plot all chromosomes at once}

\item{chr_id}{the number of chromosome want to plot}

\item{species}{which species in the input data, unused at the moment}

\item{start_position}{decimal digit, default unit is 'Mb'. such as 23.2112}

\item{end_position}{decimal digit, default unit is 'Mb'. such as 23.2112}

\item{individual_id}{the ID of individual in cnv list, used to plot all chromosome for specific Individual}

\item{max_chr_length}{the maximum length of chromosome, default unit is 'Mb'}

\item{plot_gene}{if true, will plot gene above the CNV plot, need to combine with the clean_cnv, and the stardard clean_cnv file need the annotated cnv list which was generated from call_gene function}

\item{refgene}{internal reference gene-list are "ARS_ens", "ARS_UCSC" and "UMD_UCSC". Or provide the reference genes list corresponding to your data}

\item{plot_title}{set the title of final plot}

\item{report_id}{report the sample ID while plotting}

\item{pedigree}{pedigree list, require at least three columns, Sample_Id, Sire and Dam}

\item{show_name}{default value is show_name = c(0, 160). accept the vectors only, unit is Mb. for example show_name = c(11.2, 12.4, 15.3, 18.4), means only plot the genes within the given interval}

\item{width_1}{number to set the width of final plot size, unit is 'cm'}

\item{height_1}{number to set the height of final plot size, unit is 'cm'}

\item{folder}{set the name of folder to save results}
}
\value{
CNV distribution plot
}
\description{
Title visual CNV
Visualizing CNV on both population level and individual level
}
\examples{

}
