% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvr_visualising.R
\name{cnvr_plot}
\alias{cnvr_plot}
\title{Title}
\usage{
cnvr_plot(
  cnvr,
  assembly = "ARS",
  cnv_annotation = NULL,
  sample_size = NULL,
  common_cnv_threshold = 0.05
)
}
\arguments{
\item{common_cnv_threshold}{}
}
\value{

}
\description{
Title
}
