% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cnv.R
\name{compare_cnv}
\alias{compare_cnv}
\title{Title compare_cnv
The idea to compare CNV between ars and umd are find out how many differents are there
we defined 6 comparison standards of CNV
1) overlaped
1.same start and end, same SNP inside, fully overlap
2.same start and end, diffrent snp number, fully overlap
3.different start or end, overlaped, partial overlap
2) non-overlap
4.missing start or end position
5.End <= start
6.different start or end, non-overlap
according to the codintion, the first thing is to match coordinates for both version
then find overlap cnv and non overlapcnv
then summarize how many CNVs are in above standards}
\usage{
compare_cnv(cnv_umd, cnv_ars, umd_ars_map = NULL, width_1 = 14, height_1 = 11)
}
\value{

}
\description{
Title compare_cnv
The idea to compare CNV between ars and umd are find out how many differents are there
we defined 6 comparison standards of CNV
1) overlaped
1.same start and end, same SNP inside, fully overlap
2.same start and end, diffrent snp number, fully overlap
3.different start or end, overlaped, partial overlap
2) non-overlap
4.missing start or end position
5.End <= start
6.different start or end, non-overlap
according to the codintion, the first thing is to match coordinates for both version
then find overlap cnv and non overlapcnv
then summarize how many CNVs are in above standards
}
