% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cnvr.R
\name{compare_cnvr}
\alias{compare_cnvr}
\title{Title compare_cnvr
#The idea to compare CNV between ars and umd are find out how many differents are there}
\usage{
compare_cnvr(cnvr_umd, cnvr_ars, umd_ars_map = NULL)
}
\arguments{
\item{umd_ars_map}{}
}
\value{

}
\description{
Title compare_cnvr
#The idea to compare CNV between ars and umd are find out how many differents are there
}
