% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coord_cnvr.R
\name{compare_interval}
\alias{compare_interval}
\title{Title compare_interval
compare cnvr, each input file should contain Chr, Start and End columns,
the Chr column should be the number only, for example: 1 not chr1}
\usage{
compare_interval(interval_1, interval_2, map = NULL)
}
\arguments{
\item{map}{}
}
\value{
overlap
}
\description{
Title compare_interval
compare cnvr, each input file should contain Chr, Start and End columns,
the Chr column should be the number only, for example: 1 not chr1
}
