% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coord.R
\name{convert_coord}
\alias{convert_coord}
\title{Convert coordinates of genomic intervals}
\usage{
convert_coord(
  input_tar = NULL,
  input_def = NULL,
  def_tar_map,
  proxy_range = 20000,
  len_diff_threshold = 1e+05,
  folder = "convert_coord"
)
}
\arguments{
\item{input_tar}{the interval list, the version of reference genome should be the same as one in the 'def_tar_map'.}

\item{input_def}{the interval list  the version of reference genome should be the same as one in the 'def_tar_map'.}

\item{def_tar_map}{map file contains coordinates in both version of map. default file was generated by the 'convert_map' function.}

\item{proxy_range}{the range to match the nearest proxy position to missing location SNP}

\item{len_diff_threshold}{the maximum length difference allowed between the converted Intervals and original intervals}

\item{folder}{create new folder to save results}
}
\value{
Interval with converted map
}
\description{
The input file should have at least three columns: Chr, Start and End.
}
\details{
Quality control Principle of conversion
1.End must larger than Start, i.e. the length of the interval must be a positive value
2.Both the Start and End position must match rows in the def_tar_map file
3.The difference of length between Converted Interval and Original Intervals should be less than threshold `len_diff_threshold`
4.If allow to replace the missing points by nearest point, new interval should fit all above conditions. For Start position, matching proxy from down stream, for End position, matching proxy from up stream
}
