% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_map_coord.R
\name{convert_map}
\alias{convert_map}
\title{Title Prepare SNP Map}
\usage{
convert_map(umd_map, standard_ARS_map)
}
\arguments{
\item{convert_map}{}
}
\value{

}
\description{
Title Prepare SNP Map
}
