% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_interval.R
\name{generate_interval_decimal}
\alias{generate_interval_decimal}
\title{Title generate_interval_numeric
this function is used to generate interval with decimal
function to generate all intervals by random combination of two number from the given two number
the intervals must has order, the start number must smaller than the end
generate all non-duplicated intervals combination}
\usage{
generate_interval_decimal(i, j)
}
\arguments{
\item{i}{the start value, it must be A Decimal number, such as 0.1}

\item{j}{the end value, it must be A Decimal number, such sa 0.2}
}
\value{
interval
}
\description{
Title generate_interval_numeric
this function is used to generate interval with decimal
function to generate all intervals by random combination of two number from the given two number
the intervals must has order, the start number must smaller than the end
generate all non-duplicated intervals combination
}
