% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_interval.R
\name{generate_interval_integer}
\alias{generate_interval_integer}
\title{Title generate_interval_integer
function to generate all intervals by random combination of two number from the given two number
the intervals must has order, the start number must smaller than the end
generate all non-duplicated intervals combination}
\usage{
generate_interval_integer(i, j)
}
\arguments{
\item{i}{the start number, must be integer, such as 1}

\item{j}{the end number, must be integer, such as 10}
}
\value{
interval
}
\description{
Title generate_interval_integer
function to generate all intervals by random combination of two number from the given two number
the intervals must has order, the start number must smaller than the end
generate all non-duplicated intervals combination
}
