% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dome_data.R
\name{get_demo}
\alias{get_demo}
\title{Title get_demo
The purpose of this function is to prepare demo data for 'HandyCNV'. The main reason is because of the SNP Signal Intensity and Genotype files always too large,
when we making plots can select the data inside all CNVR regions instead of using the whole file. The output will extract all Signal Intensity, genotype and map
information of given CNVRs. The new generated ped and map files keeps the same format with Plink Software, so we can --make-bed or --recode in Plink as well.}
\usage{
get_demo(
  intensity = NULL,
  ped = NULL,
  map = NULL,
  cnvr = NULL,
  freq_threshold = 0,
  folder = "demo_data"
)
}
\arguments{
\item{intensity}{Four columns by order with}

\item{ped}{Export from GenomeStudio}

\item{map}{Export from GenomeStudio, requires four columns without header.}

\item{cnvr}{Generate by 'call_cnvr', requires at least have four columns: Chr, Start, End, Frequency}

\item{freq_threshold}{This used to filter the CNVRs, only prepare the demo data for CNVRs which passed this threshold}

\item{folder}{Set name of new folder to save results}
}
\value{
Demo data of Intensity, Map, Ped file which appears only in CNVRs
}
\description{
Title get_demo
The purpose of this function is to prepare demo data for 'HandyCNV'. The main reason is because of the SNP Signal Intensity and Genotype files always too large,
when we making plots can select the data inside all CNVR regions instead of using the whole file. The output will extract all Signal Intensity, genotype and map
information of given CNVRs. The new generated ped and map files keeps the same format with Plink Software, so we can --make-bed or --recode in Plink as well.
}
