% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refgene.R
\name{get_refgene}
\alias{get_refgene}
\title{Title get_refgene
Download and make standard format of reference gene from UCSC website}
\usage{
get_refgene(gene_version = NULL)
}
\arguments{
\item{gene_version}{The version of reference gene 'HandyCNV' support to download, if now don't have the version of the gene you want, please feel free to cotact the Maintainer to help you updating this function.}
}
\value{
Standard formatted reference gene list
}
\description{
Title get_refgene
Download and make standard format of reference gene from UCSC website
}
