% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_haplotype.R
\name{haplo_visual}
\alias{haplo_visual}
\title{The order of haplotype present in plot are start from high frequency to low from top to bottom}
\usage{
haplo_visual(
  haplotype,
  filter_freq = 0,
  letter_size = 2,
  show_letter = TRUE,
  xlab_text = ""
)
}
\arguments{
\item{haplotype}{a list of haplotype information that generate by get_haplotype function}

\item{filter_freq}{used to filter haplotype by frequency, only the selected haplotype will present in plot}

\item{letter_size}{integer number to adjust the letter size of GCTA etc}

\item{show_letter}{assign "True" to present base pairs in plot, "False" to hide text, is useful on large plot}

\item{xlab_text}{add additional text into X label, for example, plot haplotype for specific genes}
}
\value{
haplotype plot
}
\description{
The order of haplotype present in plot are start from high frequency to low from top to bottom
}
