% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv_inten_maf.R
\name{plot_cnv_inten_maf}
\alias{plot_cnv_inten_maf}
\title{Title}
\usage{
plot_cnv_inten_maf(
  cnvr,
  cnv_annotation,
  intensity,
  map,
  prefix_bed,
  sample_size,
  common_cnv_threshold = 0.05
)
}
\arguments{
\item{cnvr}{cnvr list, standard file was generated from call_cnvr function}

\item{cnv_annotation}{standard file was generaed from call_gene function}

\item{intensity}{the signal inyensity file contains LRR and BAF}

\item{map}{the map corresponding to the reference genome in cnvr file, standard file was generated from convert_map function}

\item{prefix_bed}{the prefix of bed, bim and fam file in plink format}

\item{sample_size}{the total number of unique samples in cnvr list}

\item{common_cnv_threshold}{two decimal places, combine with sample_size to set the common threshold}
}
\value{

}
\description{
Title
}
