% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_source_plot.R
\name{plot_cnv_source}
\alias{plot_cnv_source}
\title{Title plot_cnv_source}
\usage{
plot_cnv_source(cnvr, clean_cnv, pedigree, Frequent_threshold)
}
\arguments{
\item{cnvr}{standard file was generated by 'call_cnvr' function, should contain the frequency of CNVR}

\item{clean_cnv}{standard file was generated by 'cnv_clean' function}

\item{pedigree}{include 'Sample_ID', 'Father_ID', 'Mother_ID', 'Country', 'Group'}

\item{Frequent_threshold}{Integer, indicate the frequency of CNVR}
}
\value{
CNVR source plot
}
\description{
Title plot_cnv_source
}
