% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_visual.R
\name{plot_gene}
\alias{plot_gene}
\title{Custom plotting gene from reference gene list}
\usage{
plot_gene(
  refgene = NULL,
  chr_id,
  start,
  end,
  show_name = c(0, 160),
  cnv = NULL,
  height_1 = 2,
  width_1 = 10,
  col_gene = "gray",
  gene_font_size = 2.5
)
}
\arguments{
\item{refgene}{reference gene list. Standard input file was generated by 'get_refgene' function.}

\item{chr_id}{chromosome ID, should be the integer type only, such as 1 indicates Chromosome 1.}

\item{start}{the start physical position used in the plot, the unit is Mb.}

\item{end}{the end physical position used in plot, unit in Mb.}

\item{show_name}{the default value of show_name = c(0, 160). accept the vectors only, unit in Mb. for example show_name = c(11.2, 12.4, 15.3, 18.4), means only plot the genes within the given interval}

\item{cnv}{is support for both CNV and ROH, set cnv argument as TRUE will plot gene for CNV, otherwise will plot gene for ROH
interval 11.2-12.4 Mb and 15.3-18.4 Mb, the maximum pairs of intervals are three.}

\item{height_1}{set height of gene plot.}

\item{width_1}{set width of gene plot.}

\item{col_gene}{set the color of Gene, only work when the 'cnv' argument was set as 'TRUE'.}

\item{gene_font_size}{set the size of gene font}
}
\value{
gene plot with given interval
}
\description{
The function is used for plotting gene from reference gene list, the standard reference gene lists was generated by 'get_refgene' function in HandyCNV.
}
