% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity.R
\name{plot_intensity}
\alias{plot_intensity}
\title{Title plot intensity of cnvr}
\usage{
plot_intensity(
  cnvr,
  cnv_annotation,
  intensity,
  map,
  prefix_bed,
  sample_size,
  common_cnv_threshold = 0.05,
  chr_id,
  start_position,
  end_position
)
}
\arguments{
\item{cnvr}{cnvr list, standard file was generated from call_cnvr function}

\item{cnv_annotation}{standard file was generated from call_gene function}

\item{intensity}{the signal intensity file contains LRR and BAF}

\item{map}{the map corresponding to the reference genome in cnvr file, standard file was generated from convert_map function}

\item{prefix_bed}{the prefix of bed, bim and fam file in plink format}

\item{sample_size}{the total number of unique samples in cnvr list}

\item{common_cnv_threshold}{two decimal places, combine with sample_size to set the common threshold}

\item{chr_id}{the number of chromosome, integer, such as chromosome 1 is: 1}

\item{start_position}{decimal digit, default unit is 'Mb'. such as 23.2112}

\item{end_position}{decimal digit, default unit is 'Mb'. such as 23.2112}
}
\value{
CNVR details plot
}
\description{
Title plot intensity of cnvr
}
