% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_haplotype.R
\name{prep_phased}
\alias{prep_phased}
\title{Title prep_phased
This function used to prepare phased genotype file to standard format in 'HandyCNV' to get haplotype for any interested genomic region,
The input file should phased by Beagle 5.1.}
\usage{
prep_phased(phased_geno)
}
\arguments{
\item{phased_geno}{phased genotype file from Beagle 5.1, the first nine columns contain basic information, the rest are genotype}
}
\value{
A list contain map and genotype information
}
\description{
Title prep_phased
This function used to prepare phased genotype file to standard format in 'HandyCNV' to get haplotype for any interested genomic region,
The input file should phased by Beagle 5.1.
}
