% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_visual.R
\name{roh_visual}
\alias{roh_visual}
\title{Title
Visualizing ROH on both population level and individual level}
\usage{
roh_visual(
  clean_roh,
  max_chr = NULL,
  chr_id = NULL,
  chr_length = NULL,
  start_position = NULL,
  end_position = NULL,
  individual_id = NULL,
  gene = NULL,
  plot_title = NULL,
  report_id = NULL,
  pedigree = NULL,
  show_name = c(0, 160),
  width_1 = 13,
  height_1 = 10
)
}
\arguments{
\item{chr_id}{the number of chromosome want to plot}

\item{chr_length}{set the length for the chromosome}

\item{individual_id}{reproting ID of individuals while plotting}

\item{plot_title}{add title in the polt}

\item{roh}{roh results from plink or CNVPartition}

\item{max_chr_length}{the maximun length of chromosome in the aimed species, it used to set the boundary of the polt}

\item{plot_gene}{present genes in given interval}
}
\value{
ROH distribution plot
}
\description{
Title
Visualizing ROH on both population level and individual level
}
\examples{

}
