% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_cnvr.R
\name{call_cnvr}
\alias{call_cnvr}
\title{Call CNV Regions}
\usage{
call_cnvr(clean_cnv, roh = NULL, chr_set = 29, folder = "UMD")
}
\arguments{
\item{clean_cnv}{a clean CNV file that was generated by the clean_cnv function}

\item{roh}{ROH file from the clean_cnv function; only CNVPartition results will generate ROH results}

\item{chr_set}{set the maximum number of chromosomes in the CNV list}

\item{folder}{set the name of the output folder.}
}
\value{
Four or five tables: CNVR list, brief summary, and individual CNV and CNVR ID.
}
\description{
Format CNV results from PennCNV and CNVPartition to the standard format
Call CNVR, summary individual CNVR type, report CNVR frequency and type
}
