% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvr_annotation.R
\name{call_gene}
\alias{call_gene}
\title{Title call_gene}
\usage{
call_gene(
  refgene = system.file("extdata", "Demo_data/gene_annotation/ensGene_ars_210202.txt",
    package = "HandyCNV"),
  interval = NULL,
  clean_cnv = NULL
)
}
\arguments{
\item{refgene}{The reference genes list corresponding to your data}

\item{interval}{Could be CNV, ROH, QTL or any interval list. At least comprised by four columns Interval_ID, Chr, Start and End}

\item{clean_cnv}{The output data from cnv_clean function}
}
\value{

}
\description{
This function used for annotating genes for any interval by giving the interval list and reference gene list.
The interval file requirs at least consists of four cloumns, they are ID, Chr, Start and End.
The first column must be the ID column, the column of Chr should only contain the number of chromosome, the units of Start and End columns are the basepair.
}
