% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refgene.R
\name{clean_ensgene}
\alias{clean_ensgene}
\title{Title clean_ensgene
clean refgene for ensembl gene.
Main works are add column names for table, add gene names to new table, replace missing gene names
with Ensembl ID then remove duplicated gene, then return the clean formatted gene list.}
\usage{
clean_ensgene(refgene, ens_id)
}
\arguments{
\item{refgene}{ensembl reference gene download from UCSC website}

\item{ens_id}{ensembl id download from UCSC website}
}
\value{
formatted gene list.
}
\description{
Title clean_ensgene
clean refgene for ensembl gene.
Main works are add column names for table, add gene names to new table, replace missing gene names
with Ensembl ID then remove duplicated gene, then return the clean formatted gene list.
}
