% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refgene.R
\name{clean_ucsc}
\alias{clean_ucsc}
\title{Title clean_ucsc
clean refgene for UCSC gene}
\usage{
clean_ucsc(refgene)
}
\arguments{
\item{refgene}{reference gene list with UCSC version which download from UCSC website}
}
\value{
formatted gene list.
}
\description{
Title clean_ucsc
clean refgene for UCSC gene
}
