% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_haplotype.R
\name{closer_snp}
\alias{closer_snp}
\title{Pick out the names and positions of start and end SNPs of a genomic interval}
\usage{
closer_snp(phased_input, chr = 14, start, end)
}
\arguments{
\item{phased_input}{map file}

\item{chr}{chromosome ID}

\item{start}{start position, unite in bp'}

\item{end}{end position, unite in 'bp'}
}
\value{
SNP position
}
\description{
Get the closest SNPs within the given interval, working on the 'map' file generated from 'pre_phased function'.
}
