% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_clean.R
\name{cnv_clean}
\alias{cnv_clean}
\title{Title  cnv_clean
Used for cleaning the default CNV results into a standard format for the further use.
Now support to read the results from PennCNV and CNVPartition}
\usage{
cnv_clean(cnvpartition = NULL, penncnv = NULL, penn_id_sep = "cnv/")
}
\arguments{
\item{penn_id_sep}{}
}
\value{

}
\description{
Title  cnv_clean
Used for cleaning the default CNV results into a standard format for the further use.
Now support to read the results from PennCNV and CNVPartition
}
