% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_clean.R
\name{cnv_clean}
\alias{cnv_clean}
\title{Title  cnv_clean
Used for cleaning the default CNV results into a standard format for the further use.
Now support to read the results from PennCNV and CNVPartition}
\usage{
cnv_clean(cnvpartition = NULL, penncnv = NULL, penn_id_sep = "cnv/")
}
\arguments{
\item{cnvpartition}{the default CNV results from CNVPartition}

\item{penncnv}{the default CNV results from PennCNV}

\item{penn_id_sep}{the sepatator in the column of Sample ID in PennCNV results, if the ID bind with the path, need to separate by this argument}
}
\value{

}
\description{
Title  cnv_clean
Used for cleaning the default CNV results into a standard format for the further use.
Now support to read the results from PennCNV and CNVPartition
}
