% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_summary.R
\name{cnv_summary_plot}
\alias{cnv_summary_plot}
\title{Title summarise CNVs
Make graphes for CNVs}
\usage{
cnv_summary_plot(clean_cnv, plot_sum_1 = NULL, plot_sum_2 = NULL)
}
\arguments{
\item{clean_cnv}{the clean cnv file was generate by clean_cnv function}

\item{plot_sum_1}{first type of combination of summary plot}

\item{plot_sum_2}{second type of combination of summary plot}
}
\value{

}
\description{
Title summarise CNVs
Make graphes for CNVs
}
