% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_visualising.R
\name{cnv_visual}
\alias{cnv_visual}
\title{Title
Visualizing CNV on both population level and individual level}
\usage{
cnv_visual(
  clean_cnv,
  max_chr = NULL,
  chr_id = NULL,
  chr_length = NULL,
  start_position = NULL,
  end_position = NULL,
  individual_id = NULL,
  max_chr_length = 160,
  plot_gene = NULL,
  plot_title = NULL,
  report_id = NULL,
  pedigree = NULL,
  show_name = c(0, 160),
  width_1 = 13,
  height_1 = 10
)
}
\arguments{
\item{plot_title}{}
}
\value{
CNV distribution plot
}
\description{
Title
Visualizing CNV on both population level and individual level
}
\examples{

}
