% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvr_visualising.R
\name{cnvr_plot}
\alias{cnvr_plot}
\title{Title}
\usage{
cnvr_plot(
  cnvr,
  assembly = "ARS",
  cnv_annotation = NULL,
  sample_size = NULL,
  common_cnv_threshold = 0.05
)
}
\arguments{
\item{cnvr}{cnvr file generated from call_cnvr function}

\item{assembly}{which reference genome assenbly used in the data, each chromosome has different length between various assembly, now only support Bovine UMD3.1 and ARS1.2}

\item{cnv_annotation}{gene annotated cnv file, generated from call_gene function}

\item{sample_size}{integer, the total number of unique samples in the cnv result. combine with common_cnv_threshold to plot all CNVs which passed the threshold}

\item{common_cnv_threshold}{two decimal places, cobine with sample_size to plot all CNVs passed the common threshold}
}
\value{

}
\description{
Title
}
