% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvr_plot.R
\name{cnvr_plot}
\alias{cnvr_plot}
\title{Title cnvr_plot}
\usage{
cnvr_plot(
  cnvr,
  assembly = "ARS",
  legend_x = 127,
  legend_y = 30,
  clean_cnv = NULL,
  sample_size = NULL,
  common_cnv_threshold = 0.05,
  refgene = "ARS",
  gain_col = "red",
  loss_col = "green",
  mixed_col = "blue",
  folder = "cnvr_plot"
)
}
\arguments{
\item{cnvr}{cnvr file generated from call_cnvr function}

\item{assembly}{which reference genome assembly used in the data, each chromosome has different length between various assembly, now only support Bovine UMD3.1 and ARS1.2}

\item{legend_x}{the x coordinate of legend, relative to the maximum length of Chromosome, unit is 'Mb'}

\item{legend_y}{the y coordinate of legend}

\item{clean_cnv}{standard input file was generated by 'cnv_clean' function}

\item{sample_size}{integer, the total number of unique samples in the cnv result. combine with common_cnv_threshold to plot all CNVs which passed the threshold}

\item{common_cnv_threshold}{two decimal places, combine with sample_size to plot all CNVs passed the common threshold}

\item{refgene}{reference gene list, use for plot gene}

\item{gain_col}{set color for type of gain CNVR}

\item{loss_col}{set color for type of loss CNVR}

\item{mixed_col}{set colof for type of mixed CNVR}

\item{folder}{set name of folder to save results}
}
\value{
A figure of CNVR distribution map and plot parameters.
If given clean_cnv file, will plot all CNVRs which are passed common threshold.
}
\description{
Title cnvr_plot
}
