% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvr_plot.R
\name{cnvr_plot}
\alias{cnvr_plot}
\title{Custom CNVR distribution map and plot all high frequency CNVR at once}
\usage{
cnvr_plot(
  cnvr,
  assembly = "ARS",
  overlap_cnvr = NULL,
  label_prop = TRUE,
  left_prop_label = 3,
  legend_x = 127,
  legend_y = 30,
  clean_cnv = NULL,
  sample_size = NULL,
  common_cnv_threshold = 0.05,
  refgene = "ARS",
  gain_col = "red",
  loss_col = "green",
  mixed_col = "blue",
  overlap_col = "purple",
  chr_col = "black",
  folder = "cnvr_plot",
  gene_font_size = 2.2,
  width_1 = 22,
  height_1 = 14.5,
  col_gene = "gray"
)
}
\arguments{
\item{cnvr}{standard cnvr file was generated by 'call_cnvr' function}

\item{assembly}{UMD = "Bovine UMD 3.1", ARS = "BOvine ARS UCD 1.2", "hg38" = "Human hg38", "hg19" = "Human hg19", Sheep = "Oar_v4.0", Pig = "Sscrofa11.1", Chicken = "Chicken_galGal6", Horse = "EquCab3.0", Dog = "UMICH_Zoey_3.1"}

\item{overlap_cnvr}{the common CNVRs list between two CNVR results, which use to mark underline to indicate the common CNVRs, should only contain Chr, Start and End three columns, the standard file was generated by 'compare_cnvr' function}

\item{label_prop}{if 'TRUE', will display the proportion of CNVRs in each chromosome}

\item{left_prop_label}{adjust the coordinates of the label to move to the left, default value is 3}

\item{legend_x}{the x coordinate of legend, relative to the maximum length of Chromosome, unit is 'Mb'}

\item{legend_y}{the y coordinate of legend}

\item{clean_cnv}{standard input file was generated by 'cnv_clean' function}

\item{sample_size}{integer, the total number of unique samples in the cnv result. combine with common_cnv_threshold to plot all CNVs which passed the threshold}

\item{common_cnv_threshold}{two decimal places, combine with sample_size to plot all CNVs passed the common threshold}

\item{refgene}{reference gene list, use for plot gene}

\item{gain_col}{set color for type of gain CNVR}

\item{loss_col}{set color for type of loss CNVR}

\item{mixed_col}{set color for type of mixed CNVR}

\item{overlap_col}{set color for Overlapped CNVRs}

\item{chr_col}{set color of Chromosomes}

\item{folder}{set name of folder to save results}

\item{gene_font_size}{set the font size of gene in CNV annotation plot}

\item{width_1}{number to set the width of final plot size, unit is 'cm'}

\item{height_1}{number to set the height of final plot size, unit is 'cm'}

\item{col_gene}{set the color of Gene, only work in Gene annotated at CNV plot}
}
\value{
A figure of CNVR distribution map and plot parameters.
If given clean_cnv file, will plot all CNVRs which are passed common threshold.
}
\description{
Custom CNVR distribution map and plot all high frequency CNVR at once
}
