% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cnv.R
\name{compare_cnv}
\alias{compare_cnv}
\title{Compare  CNV}
\usage{
compare_cnv(
  cnv_def,
  cnv_tar,
  def_tar_map = NULL,
  width_1 = 14,
  height_1 = 11,
  legend_x = 0.9,
  legend_y = 0.9,
  folder = "compare_cnv",
  col_1 = "pink",
  col_2 = "lightblue",
  plot_caption = TRUE
)
}
\arguments{
\item{cnv_def}{'def' is 'default' indicate the first CNV input file, the standard file was generated by 'clean_cnv' function}

\item{cnv_tar}{'tar' is 'target' indicate the second CNV input file, the standard file was generated by 'clean_cnv' function}

\item{def_tar_map}{map file contains coordinates in both version of map. only need in comparison between the results from different versions. standard file was generated by 'convert_map' function}

\item{width_1}{integer, default value is 14, set the width of final plot size, unit is 'cm'}

\item{height_1}{integer, default value is 11, set the height of final plot size, unit is 'cm'}

\item{legend_x}{decimal digit, default value is 0.9, consistent with ggplot manual coordinates of legend}

\item{legend_y}{decimal digit, default value is 0.9, consistent with ggplot manual coordinates of legend}

\item{folder}{set name of folder to save results}

\item{col_1}{set color for non-overlapped bar}

\item{col_2}{set color for overlapped bar}

\item{plot_caption}{If TRUE, present Note Caption in Comparison plot}
}
\value{
Details comparison results of CNVs between input lists.
}
\description{
The idea to compare CNVs between different results are find out how many different are there in both individual and population level.
For CNVs we more concern about the number and proportion of overlapped CNVs.
}
\details{
Six comparison standards of CNV were defined:
1) overlapped
1.same start and end, same SNP inside, fully overlap
2.same start and end, different snp number, fully overlap
3.different start or end, overlapped, partial overlap
2) non-overlap
4.missing start or end position
5.End <= start
6.different start or end, non-overlap
according to the condition, the first thing is to match coordinates for both version
then find overlap cnv and non overlap cnv
then summarize how many CNVs are in above standards
}
