% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cnv.R
\name{compare_cnv}
\alias{compare_cnv}
\title{Title compare_cnv
The idea to compare CNV between ars and umd are find out how many differents are there
we defined 6 comparison standards of CNV
1) overlaped
1.same start and end, same SNP inside, fully overlap
2.same start and end, diffrent snp number, fully overlap
3.different start or end, overlaped, partial overlap
2) non-overlap
4.missing start or end position
5.End <= start
6.different start or end, non-overlap
according to the codintion, the first thing is to match coordinates for both version
then find overlap cnv and non overlapcnv
then summarize how many CNVs are in above standards}
\usage{
compare_cnv(cnv_umd, cnv_ars, umd_ars_map = NULL, width_1 = 14, height_1 = 11)
}
\arguments{
\item{cnv_umd}{first cnv list, not limit on umd or ars, defaults file is the result from clean_cnv function}

\item{cnv_ars}{second cnv list, not limit on umd or ars, defaults file is the result from clean_cnv function}

\item{umd_ars_map}{map file contains coordinats in both version of map. only need in comparison between the results from different versions. default file is generated from convert_map function}

\item{width_1}{number to set the width of final plot size, unit is 'cm'}

\item{height_1}{number to set the height of final plot size, unit is 'cm'}
}
\value{

}
\description{
Title compare_cnv
The idea to compare CNV between ars and umd are find out how many differents are there
we defined 6 comparison standards of CNV
1) overlaped
1.same start and end, same SNP inside, fully overlap
2.same start and end, diffrent snp number, fully overlap
3.different start or end, overlaped, partial overlap
2) non-overlap
4.missing start or end position
5.End <= start
6.different start or end, non-overlap
according to the codintion, the first thing is to match coordinates for both version
then find overlap cnv and non overlapcnv
then summarize how many CNVs are in above standards
}
