% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cnvr.R
\name{compare_cnvr}
\alias{compare_cnvr}
\title{Compare CNVR}
\usage{
compare_cnvr(
  cnvr_def,
  cnvr_tar,
  def_tar_map = NULL,
  width_1 = 15,
  height_1 = 15,
  hjust_prop = 0,
  hjust_num = 1.5,
  folder = "compare_cnvr",
  col_1 = "gray20",
  col_2 = "springgreen4"
)
}
\arguments{
\item{cnvr_def}{'def' is 'default' indicate the first input file, standard input file was generated by 'call_cnvr' function.}

\item{cnvr_tar}{'tar' is 'target' indicate the second input file, standard input file was generated by 'call_cnvr' function.}

\item{def_tar_map}{Unused at the moment. map file contains coordinates in both version of map. only need in comparison between the results from different versions. default file is generated from convert_map function}

\item{width_1}{set the width of final plot size, unit is 'cm'}

\item{height_1}{set the height of final plot size, unit is 'cm'}

\item{hjust_prop}{default value is 0.0. used to adjust horizontal position of the number of overlapped CNVR in the plot}

\item{hjust_num}{default value is 1.5. used to adjust horizontal position of the number of overlapped CNVR in the plot}

\item{folder}{set name of folder to save results}

\item{col_1}{set color for overlapped bar}

\item{col_2}{set color for non-overlapped bar}
}
\value{
Details comparison results of CNVRs between input lists
}
\description{
The idea of CNVRs comparison is to find out how many CNVRs and how many length are overlapped between different CNVR results.
}
\details{
The six standards are defined for CNVR comparison:
1) overlapped
1.same start and end, same SNP inside, fully overlap
2.same start and end, different snp number, fully overlap
3.different start or end, overlapped, partial overlap
2) non-overlap
4.missing start or end position
5.End <= start
6.different start or end, non-overlap
according to the condition, the first thing is to match coordinates for both version
then find overlap CNVR and non overlap CNVR
then summarize how many CNVRs are in above standards
}
