% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cnvr.R
\name{compare_cnvr}
\alias{compare_cnvr}
\title{Title compare_cnvr
#The idea to compare CNV between ars and umd are find out how many differents are there}
\usage{
compare_cnvr(
  cnvr_umd,
  cnvr_ars,
  umd_ars_map = NULL,
  width_1 = 15,
  height_1 = 15
)
}
\arguments{
\item{cnvr_umd}{first cnvr list, not limited on umd or ars, default file is the result from call_cnvr function}

\item{cnvr_ars}{second cnvr list, not limited on umd or ars, default file is the result from call_cnvr function}

\item{umd_ars_map}{map file contains coordinats in both version of map. only need in comparison between the results from different versions. default file is generated from convert_map function}

\item{width_1}{number to set the width of final plot size, unit is 'cm'}

\item{height_1}{number to set the height of final plot size, unit is 'cm'}
}
\value{

}
\description{
Title compare_cnvr
#The idea to compare CNV between ars and umd are find out how many differents are there
}
