% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cnvr.R
\name{compare_cnvr}
\alias{compare_cnvr}
\title{Title compare_cnvr
#The idea to compare CNVR between different results are find out how many differences are there}
\usage{
compare_cnvr(
  cnvr_def,
  cnvr_tar,
  def_tar_map = NULL,
  width_1 = 15,
  height_1 = 15,
  hjust_prop = 0,
  hjust_num = 1.5,
  folder = "compare_cnvr",
  col_1 = "gray20",
  col_2 = "springgreen4"
)
}
\arguments{
\item{cnvr_def}{first cnvr list, not limited on def or tar, default file is the result from call_cnvr function}

\item{cnvr_tar}{second cnvr list, not limited on def or tar, default file is the result from call_cnvr function}

\item{def_tar_map}{map file contains coordinates in both version of map. only need in comparison between the results from different versions. default file is generated from convert_map function}

\item{width_1}{number to set the width of final plot size, unit is 'cm'}

\item{height_1}{number to set the height of final plot size, unit is 'cm'}

\item{hjust_prop}{default value is 0.0. used to adjust horizontal position of the number of overlapped CNVR in the plot}

\item{hjust_num}{default value is 1.5. used to adjust horizontal position of the number of overlapped CNVR in the plot}

\item{folder}{set name of folder to save results}

\item{col_1}{set color for overlapped bar}

\item{col_2}{set color for non-overlapped bar}
}
\value{
Details comparison results of CNVRs between input lists
}
\description{
Title compare_cnvr
#The idea to compare CNVR between different results are find out how many differences are there
}
