% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gene.R
\name{compare_gene}
\alias{compare_gene}
\title{Find the consensus genes by comparing the frequency of gene in which has CNV status}
\usage{
compare_gene(
  gene_freq_1,
  gene_freq_2,
  gene_freq_3 = NULL,
  gene_freq_4 = NULL,
  common_gene_threshold = 3,
  title_1 = "list_1",
  title_2 = "list_2",
  title_3 = "list_3",
  title_4 = "list_4",
  height_1 = 10,
  width_1 = 14,
  folder = "compre_gene",
  col_1 = "red",
  col_2 = "yellow",
  col_3 = "dodgerblue",
  col_4 = "pink2"
)
}
\arguments{
\item{gene_freq_1}{first genelist, the default file was generated from call_gene function}

\item{gene_freq_2}{second genelist, the default file was generated from call_gene function}

\item{gene_freq_3}{third genelist, the default file was generated from call_gene function}

\item{gene_freq_4}{fourth genelist, the default file was generated from call_gene function}

\item{common_gene_threshold}{integer input.could calculate by multiplying The Number of Sample by 0.05 or 0.1, et al}

\item{title_1}{set title(two and three lists) or label(four lists) in figure}

\item{title_2}{set title(two and three lists) or label(four lists) in figure}

\item{title_3}{set title(two and three lists) or label(four lists) in figure}

\item{title_4}{set title(two and three lists) or label(four lists) in figure}

\item{height_1}{customize the height to save figure}

\item{width_1}{customize the width to save figure}

\item{folder}{set the name of folder to save results}

\item{col_1}{set color for common_high in two gene lists comparison plot or the color of High frequency gene in multiple comparison heatmap plot}

\item{col_2}{set color for common_low in two gene lists comparison plot or the color of Low frequency gene in multiple comparison heatmap plot}

\item{col_3}{set color for High_Freq_list_1 in two gene lists comparison plot}

\item{col_4}{set color for High_Freq_list_2 in two gene lists comparison plot}
}
\value{
Comparison summary results and plots.
}
\description{
Do comparison of genes between multiple annotated gene frequency lists. The default genelist file was generate by call_gene function.
If the users customize the genelists, it requires at least two columns in each gene list, including fixed column names name2(Gene name) and Frequency(frequency of gene)
}
