% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_comparation.R
\name{compare_gene}
\alias{compare_gene}
\title{Title Gene Comparison
There are two application scenarios:Compare two gene lists or three gene lists}
\usage{
compare_gene(
  gene_freq_1,
  gene_freq_2,
  gene_freq_3 = NULL,
  common_gene_threshold = 3,
  title_1 = "genelist_1",
  title_2 = "genelist_2",
  title_3 = "genelist_3"
)
}
\arguments{
\item{common_gene_threshold}{}
}
\value{

}
\description{
Title Gene Comparison
There are two application scenarios:Compare two gene lists or three gene lists
}
