% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coord_cnvr.R
\name{convert_coord}
\alias{convert_coord}
\title{Title convert_coord
only need Chr, Start and End}
\usage{
convert_coord(input_ars = NULL, input_umd = NULL, map)
}
\arguments{
\item{map}{}
}
\value{

}
\description{
Title convert_coord
only need Chr, Start and End
}
