% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coord_cnvr.R
\name{convert_coord}
\alias{convert_coord}
\title{Title convert_coord
only need Chr, Start and End}
\usage{
convert_coord(input_ars = NULL, input_umd = NULL, map)
}
\arguments{
\item{input_ars}{the interval list with ARS1.2 coordinates, not limited on ARS once given the right map file}

\item{input_umd}{the interval list with UMD3.1 coordinates, not limited on UMD once given the right map file}

\item{map}{map file contains coordinats in both version of map. default file is generated from convert_map function}
}
\value{

}
\description{
Title convert_coord
only need Chr, Start and End
}
