% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coord.R
\name{convert_coord}
\alias{convert_coord}
\title{Title convert_coord
only need Chr, Start and End}
\usage{
convert_coord(
  input_tar = NULL,
  input_def = NULL,
  def_tar_map,
  proxy_range = 20000,
  len_diff_threshold = 1e+05,
  folder = "convert_coord"
)
}
\arguments{
\item{input_tar}{the interval list with ARS1.2 coordinates, not limited on ARS once given the right map file}

\item{input_def}{the interval list with UMD3.1 coordinates, not limited on UMD once given the right map file}

\item{def_tar_map}{map file contains coordinates in both version of map. default file is generated from convert_map function}

\item{proxy_range}{the range to match the nearest proxy position to missing location SNP}

\item{len_diff_threshold}{the maximum length difference allowed between the converted Intervals and original intervals}

\item{folder}{create new folder to save results}
}
\value{
Interval with converted map
}
\description{
Quality control Principle of conversion
1.End must larger than Start, i.e. the length of the interval must be a positive value
2.Both the Start and End position must match rows in the def_tar_map file
3.The difference of length between Converted Interval and Original Intervals should be less than threshold `len_diff_threshold`
4.If allow to replace the missing points by nearest point, new interval should fit all above conditions. For Start position, matching proxy from down stream, for End position, matching proxy from up stream
}
