% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv_inten_maf.R
\name{plot_cnv_inten_maf}
\alias{plot_cnv_inten_maf}
\title{Title}
\usage{
plot_cnv_inten_maf(
  cnvr,
  cnv_annotation,
  intensity,
  map,
  prefix_bed,
  sample_size,
  common_cnv_threshold = 0.05
)
}
\arguments{
\item{common_cnv_threshold}{}
}
\value{

}
\description{
Title
}
