% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnvr_panorama.R
\name{plot_cnvr_panorama}
\alias{plot_cnvr_panorama}
\title{Title plot_cnvr_panorama}
\usage{
plot_cnvr_panorama(
  cnvr,
  cnv_annotation,
  intensity,
  map,
  prefix_bed,
  ld_heat = TRUE,
  sample_size,
  common_cnv_threshold = 0.05,
  width_1 = 14,
  height_1 = 30,
  folder = "cnvr_panorama"
)
}
\arguments{
\item{cnvr}{cnvr list, standard file was generated from call_cnvr function}

\item{cnv_annotation}{standard file was generated from call_gene function}

\item{intensity}{the signal intensity file contains LRR and BAF}

\item{map}{the map corresponding to the reference genome in cnvr file, standard file was generated from convert_map function}

\item{prefix_bed}{the prefix of bed, bim and fam file in plink format}

\item{ld_heat}{If TRUE, will plot LD by heatmap, If false will plotting by Classical Inverted Triangle, heatmap is fatser, triagnle may mot work well.}

\item{sample_size}{the total number of unique samples in cnvr list}

\item{common_cnv_threshold}{two decimal places, combine with sample_size to set the common threshold}

\item{width_1}{default value is 14, unit is 'cm', set the width of final plot}

\item{height_1}{default value is 30, unit is 'cm', set the height of final plot}

\item{folder}{set name of folder to save results}
}
\value{
cnvr plot with all CNVs, annotated genes, Log R Ratio, B Allele Frequency, Genotyping rate and LD...
}
\description{
Title plot_cnvr_panorama
}
