% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnvr_source.R
\name{plot_cnvr_source}
\alias{plot_cnvr_source}
\title{Plotting CNVs by the source information}
\usage{
plot_cnvr_source(
  cnvr,
  clean_cnv,
  pedigree,
  Frequent_threshold,
  height_1 = 30,
  width_1 = 15,
  folder = "plot_cnvr_source"
)
}
\arguments{
\item{cnvr}{standard file was generated by 'call_cnvr' function, should contain the frequency of CNVR}

\item{clean_cnv}{standard file was generated by 'cnv_clean' function}

\item{pedigree}{include 'Sample_ID', 'Father_ID', 'Mother_ID', 'Country', 'Group'}

\item{Frequent_threshold}{Integer, indicate the frequency of CNVR}

\item{height_1}{set the height of plot}

\item{width_1}{set the width of plot}

\item{folder}{set the folder to save results}
}
\value{
CNVR source plot
}
\description{
Plotting CNVs by the source information
}
