% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_visual.R
\name{plot_gene}
\alias{plot_gene}
\title{Title plot_gene}
\usage{
plot_gene(
  refgene = NULL,
  chr_id,
  start,
  end,
  show_name = c(0, 160),
  cnv = NULL
)
}
\arguments{
\item{refgene}{reference gene list.}

\item{chr_id}{chromosome ID, should be the integer only, like 1 or 29}

\item{start}{the start physical position used in the plot, the unit is Mb}

\item{end}{the end physical position used in plotm, the unit is Mb}

\item{show_name}{default value is show_name = c(0, 160). accept the vectors only, unit is Mb. for example show_name = c(11.2, 12.4, 15.3, 18.4), means only plot the genes within the given interval}

\item{cnv}{is support for both CNV and ROH, set cnv argument as TRUE will plot gene for CNV, otherwise will plot gene for ROH
interval 11.2-12.4 Mb and 15.3-18.4 Mb, the maximum pairs of interval are three}
}
\value{

}
\description{
Title plot_gene
}
