% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_in.R
\name{plot_inten}
\alias{plot_inten}
\title{Title}
\usage{
plot_inten(
  cnvr,
  cnv_annotation,
  intensity,
  map,
  sample_size,
  common_cnv_threshold = 0.05
)
}
\arguments{
\item{common_cnv_threshold}{}
}
\value{

}
\description{
Title
}
