% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_haplotype.R
\name{prep_phased}
\alias{prep_phased}
\title{Prepare the phased genotype into the standard format}
\usage{
prep_phased(phased_geno, convert_letter = FALSE)
}
\arguments{
\item{phased_geno}{phased genotype file from Beagle 5.1, the first nine columns contain basic information, the rest are genotype}

\item{convert_letter}{convert allele from 0, 1 to letter in REF and ALF}
}
\value{
A list contain map and genotype information
}
\description{
This function used to prepare phased genotype file to standard format in 'HandyCNV' to get haplotype for any interested genomic region,
The input file should phased by Beagle 5.1.
}
