% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_visual.R
\name{roh_visual}
\alias{roh_visual}
\title{Title
Visualizing ROH on both population level and individual level}
\usage{
roh_visual(
  clean_roh,
  max_chr = NULL,
  chr_id = NULL,
  chr_length = NULL,
  start_position = NULL,
  end_position = NULL,
  individual_id = NULL,
  gene = NULL,
  plot_title = NULL,
  report_id = NULL,
  pedigree = NULL,
  show_name = c(0, 160),
  width_1 = 13,
  height_1 = 10
)
}
\arguments{
\item{plot_title}{}
}
\value{
ROH distribution plot
}
\description{
Title
Visualizing ROH on both population level and individual level
}
\examples{

}
