% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_visual.R
\name{roh_visual}
\alias{roh_visual}
\title{Title
Visualizing ROH on both population level and individual level}
\usage{
roh_visual(
  clean_roh,
  max_chr = NULL,
  chr_id = NULL,
  chr_length = NULL,
  start_position = NULL,
  end_position = NULL,
  individual_id = NULL,
  refgene = NULL,
  plot_title = NULL,
  report_id = NULL,
  pedigree = NULL,
  show_name = c(0, 160),
  width_1 = 13,
  height_1 = 10,
  col_short = "deepskyblue",
  col_mid = "black",
  col_long = "deeppink2",
  folder = "roh_visual",
  gene_font_size = 2.2
)
}
\arguments{
\item{clean_roh}{support roh results from Plink or generated by cnv_clean form CNVPartition results}

\item{max_chr}{the maximum number of chromosomes to plot, it used for plot all chromosomes at once}

\item{chr_id}{the number of chromosome want to plot}

\item{chr_length}{set the length for the chromosome}

\item{start_position}{decimal digit, default unit is 'Mb'. such as 23.2112}

\item{end_position}{decimal digit, default unit is 'Mb'. such as 23.2112}

\item{individual_id}{reporting ID of individuals while plotting}

\item{refgene}{if true, will plot gene above roh plot}

\item{plot_title}{add title in the plot}

\item{report_id}{report the sample ID while plotting}

\item{pedigree}{pedigree list, require at least three columns, Sample_Id, Sire and Dam}

\item{show_name}{default value is show_name = c(0, 160). accept the vectors only, unit is Mb. for example show_name = c(11.2, 12.4, 15.3, 18.4), means only plot the genes within the given interval}

\item{width_1}{number to set the width of final plot size, unit is 'cm'}

\item{height_1}{number to set the height of final plot size, unit is 'cm'}

\item{col_short}{set color to shorter length of roh in distribution plot}

\item{col_mid}{set color to middle length of roh in distribution plot}

\item{col_long}{set color to longer length of roh in distribution plot}

\item{folder}{set new folder to save results}

\item{gene_font_size}{set the font size of gene in ROH annotation plot}
}
\value{
ROH distribution plot
}
\description{
Title
Visualizing ROH on both population level and individual level
}
