% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_window.R
\name{roh_window}
\alias{roh_window}
\title{Title window_size}
\usage{
roh_window(roh, window_size = 5, length_autosomal = 2489.386, threshold = 0.3)
}
\arguments{
\item{threshold}{}
}
\value{

}
\description{
Title window_size
}
