% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_window.R
\name{roh_window}
\alias{roh_window}
\title{Title window_size}
\usage{
roh_window(roh, window_size = 5, length_autosomal = 2489.386, threshold = 0.3)
}
\arguments{
\item{roh}{roh file from plink results, or generated by cnv_clean for CNVPartition results}

\item{window_size}{the length of sliding windows for deviding the chromosome, the unit is Mb}

\item{length_autosomal}{total autosomal length of the corresponding species}

\item{threshold}{the proportion used in the roh windows ditribution plot. For example 1000 samples, if threshold set as 0.5,
and 1000*0.5 = 500, the value of histgram above 500 will be color as red and the group below 500 will filling with other color}
}
\value{

}
\description{
Title window_size
}
