% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh_window.R
\name{roh_window}
\alias{roh_window}
\title{Title window_size}
\usage{
roh_window(
  roh,
  window_size = 5,
  length_autosomal = 2489.386,
  threshold = 0.3,
  folder = "roh_window",
  col_higher = "red",
  col_lower = "black",
  height_1 = 18,
  width_1 = 21,
  legend_x = 0.9,
  legend_y = 0.1,
  ncol_1 = 6
)
}
\arguments{
\item{roh}{roh file from plink results, or generated by cnv_clean for CNVPartition results}

\item{window_size}{the length of sliding windows for dividing the chromosome, the unit is Mb}

\item{length_autosomal}{Important, total autosomal length of the corresponding species, used in calculated the F_ROH}

\item{threshold}{the proportion used in the roh windows distribution plot. For example 1000 samples, if threshold set as 0.5,
and 1000*0.5 = 500, the value of histogram above 500 will be color as red and the group below 500 will filling with other color}

\item{folder}{set name of new created folder}

\item{col_higher}{set the color of bar which frequency higher than threshold in high frequent ROH plot}

\item{col_lower}{set the color of bar which frequency lower than threshold in high frequent ROH plot}

\item{height_1}{set the height of final high frequent ROH plot, units ='cm'}

\item{width_1}{set the width of final high frequent ROH plot, units = 'cm'}

\item{legend_x}{set position of legend, 0 to 1 indicate from left to right}

\item{legend_y}{set position of legend, 0 to 1 indicate from bottom to top}

\item{ncol_1}{set how many columns to present in facet plot}
}
\value{
Summary results of ROH by sliding windows across all chromosomes.
}
\description{
Title window_size
}
